Data.CAS.ContentHashable.S3

Description

Hashing of S3 objects

This module allows us to fetch objects from S3, taking advantage of S3's support for CAS to avoid the need to calculate our own content hashes.

Documentation

data ObjectInBucket obj

Reference to an object in an S3 bucket

Objects can be referenced in a few ways, so this type is parametrised over the object reference. Currently, this is expected to be: - S3.Object (alias for Text) - S3.ObjectInfo

Constructors

ObjectInBucket

Fields

_oibBucket :: Bucket

_oibObject :: obj

oibBucket :: Functor f => (Bucket -> f Bucket) -> ObjectInBucket obj -> f (ObjectInBucket obj)

A lens to _oibBucket

oibObject :: Functor f => (a -> f b) -> ObjectInBucket a -> f (ObjectInBucket b)

A lens to _oibObject

class ObjectReference a where

Methods

objectReference :: a -> Object